Support Telegram Bot API 6.1

+++ src/main/java/se/anyro/tgbotapi/types/file/Animation.java
@@ -11,5 +11,5 @@ public class Animation {
     public PhotoSize thumb;
     public String file_name;
     public String mime_type;
-    public int file_size;
+    public long file_size;
 }
